+++ packages/fs-detectors/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @vercel/fs-detectors
 
+## 6.11.2
+
+### Patch Changes
+
+- 1a1c745: [services] fix bare static service builds, allow services with only 'root' to resolve at @vercel/static
+
+## 6.11.1
+
+### Patch Changes
+
+- Updated dependencies [6b49a17]
+  - @vercel/build-utils@13.32.2
+
 ## 6.11.0
 
 ### Minor Changes