+++ packages/nestjs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @vercel/nestjs
 
+## 0.2.102
+
+### Patch Changes
+
+- @vercel/node@5.8.22
+
 ## 0.2.101
 
 ### Patch Changes